+++ px-dotnet/DataStructures/AuthorizedPayment/Payment.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MercadoPago.DataStructures.AuthorizedPayment
+{
+    public struct Payment
+    {
+        #region Properties
+
+        /// <summary>
+        /// Payment ID
+        /// </summary>
+        public long Id { get; set; }
+
+        /// <summary>
+        /// Payment status
+        /// </summary>
+        public string Status { get; set; }
+
+        /// <summary>
+        /// Further detail of the current status
+        /// </summary>
+        public string StatusDetail { get; set; }
+
+        #endregion
+    }
+}