+++ px-dotnet/SDK.cs
@@ -17,7 +17,8 @@ namespace MercadoPago
         private static string UserToken = null;
         private static int _requestsTimeout = DEFAULT_REQUESTS_TIMEOUT;
         private static int _requestsRetries = DEFAULT_REQUESTS_RETRIES;
-        private static IWebProxy _proxy;
+        private static IWebProxy _proxy;        
+        private static readonly object AccessTokenLock = new object();
         public static string RefreshToken = null;
 
         /// <summary>  
@@ -250,11 +251,15 @@ namespace MercadoPago
         /// Authenticate with MercadoPago API, using SDK.ClientId and SDK.ClientSecret. The resulting token is stored in the SDK.AccessToken property.
         /// </summary>
         /// <returns>A valid access token.</returns>
-        public static string GetAccessToken() 
-        {
-            if (String.IsNullOrEmpty(AccessToken))
+        public static string GetAccessToken()
+        {            
+            if (string.IsNullOrEmpty(AccessToken))
             {
-                AccessToken = MPCredentials.GetAccessToken(ClientId, ClientSecret);
+                lock(AccessTokenLock)
+                {
+                    if (string.IsNullOrEmpty(AccessToken))
+                        AccessToken = MPCredentials.GetAccessToken(ClientId, ClientSecret);
+                }
             }
             return AccessToken;
         }